pyproject.toml
[project]
name = "hackaton-projects"
version = "0.1.0"
description = ""
authors = [
    {name = "Germán Küber",email = "german.kuber@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "selenium (>=4.32.0,<5.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "supabase (>=2.15.1,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "termcolor (>=3.1.0,<4.0.0)",
    "torch (==2.2.2)",
    "torchvision (==0.17.2)",
    "torchaudio (==2.2.2)",
    "transformers (>=4.51.3,<5.0.0)",
    "accelerate (>=1.6.0,<2.0.0)",
    "peft (>=0.15.2,<0.16.0)",
    "datasets (>=3.6.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "numpy (<2)",
    "sentencepiece (>=0.2.0,<0.3.0)",
    "protobuf (>=6.30.2,<7.0.0)",
    "openai (>=1.77.0,<2.0.0)",
    "pinecone (>=6.0.2,<7.0.0)",
    "cohere (>=5.15.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

